requirements.txt
streamlit
pandas
langdetect
presidio-anonymizer
presidio-analyzer
python-dotenv
google-generativeai
scikit-learn
numpy
google-play-scraper
python-dateutil
groq
notion-client==2.2.1
